refactor(api)!: split error, rename globalapi and peer modules
