+++ fedimint-api-client/src/api/global_api.rs
