+++ fedimint-api-client/src/api/peer.rs
